8b8_firmware_V2.2_AY3
Firmware for AY-3-8910 chips as opposed to Yamaha YM2149, which has a built-in clock divider. Some of the original Yamaha chips are routinely retopped and rebranded as AY3, strangely enough. So if you have an 8B8 and your Clock Button isn't dividing your octave, or it makes strange(r) noises when you do, this firmware should fix the problem! Contact me at greg@semioticsounds.com with any questions.